Make cache keys unique
If there are several different queries with the same name, query caching
behavior will not know the difference as they cache key is built based
on the class name. This could potentially lead to a situation where
cached results are returned for the wrong query. To make sure that the
cache keys will stay unique, use the full name with the namespaces when
building the cache key.