Only fetch needed columns when querying all cars
Using Select(car => mapper.Map<CarDto>(car)) when converting a car
entity to a CarDto will cause all columns to be fetched from the car
table. As a lot of these columns will be discarded immediately when
mapping to a CarDto, it is better to use ProjectTo when doing to
conversion. ProjectTo will only fetch the columns that will be used in
the CarDto.